+++ project/src/components/details/details.tsx
