+++ project/src/components/favorites-empty/favorites-empty.tsx
