+++ project/src/components/favorites/favorites.tsx
